+++ R/.onLoad.r
@@ -1,0 +1,14 @@
+#' On Load Hook
+#'
+#' @description 
+#' On load hook.
+#' 
+#' @param libname 
+#' @param pkgname
+#' @template author
+#' @template references
+.onLoad <- function(libname, pkgname) {
+  
+
+  
+}