+++ rapp/apps/demo_app/options/options_runtime.r
@@ -1,5 +1,0 @@
-options_runtime <- list(
-	rapp_global = "C:/Users/rappster_admin/Documents/rapp",
-	runtime_mode = "dev",
-	lib = "Q:/home/apps/rapptools/apps/r/R-3.1.1/library"
-)