+++ rapp/apps/test/options/options_runtime.r
@@ -1,5 +1,0 @@
-options_runtime <- list(
-	rapp_global = "C:/Users/rappster_admin/Documents/rapp",
-	runtime_mode = "dev",
-	lib = .libPaths()[1]
-)