+++ rapp/apps/test/rapp/options/options.r
@@ -1,5 +1,5 @@
 options <- list(
-	ns = rapp.core.package::asPackage(x = ".")$package,
+	ns = file.path(rapp.core.package::asPackage(x = ".")$package, "test"),
 	option_2 = "your option value here (can be any R object)",
 	option_3 = "your option value here (can be any R object)"
 )