+++ tests/testthat/test-ensureProjectComponents-1.r
@@ -12,17 +12,19 @@ test_that("ensureProjectComponents", {
     path = tempdir(),
     force = TRUE
   )
+  
   path <- file.path(tempdir(), "test")
   wd_0 <- setwd(path)
   expect_true(res <- ensureProjectComponents())
   expect_true(file.exists("data"))
   expect_true(file.exists("man-roxygen"))
   expect_true(file.exists("inst/examples"))
-  expect_true(file.exists("rapp/apps/test/options"))
+  expect_true(file.exists("rapp/apps/test/rapp/options"))
   expect_true(file.exists("rapp/apps/test/R"))
   expect_true(file.exists("rapp/options"))
   expect_true(file.exists("tests/testthat/data"))
   expect_true(file.exists("vignettes"))
+  expect_true(res <- ensureProjectComponents(overwrite = TRUE))
   
   setwd(wd_0)
   on.exit({